Multi Layer ASP.NET MVC 5 sample Project with Unit Test Controller and EF6
This is one of the simple multi-layer architecture framework designed and developed by me attached is the detail technical specification for the same. 
The application built on top of MVC 5 template with maximum abstraction controller side code to interact with View from various ActionResult or JsonRresult.
 
TempData is used for caching mechanism to get the current page data with my own custom paginated List of <T>
 
EPPLUS open source codeplex dll is used to implement Export to excel functionality to export any grid data in excel template.
 
Structure map is used to implement dependency injection for email and SMS notification.
 Serialize.linq is used for passing LINQ expression to WCF layer.
Just to show the skills of WCF implementation I have implemented WCF library and WCF layer but did not consume it in Business layer. 
 Currently Business layer is directly communicates with Data model Layer.
  
Implemented Unit Test Controller class only for Visitor information Repository class 
Following are the technology components used in the solution.
1.       EF 6 Code first Approach
2.       Generic classes with some design clauses
3.       jquery.jqGrid.min.js to build grid with CRUD operation
4.       Jquery  with overriding of various functions for all common operation
5.       Bootstrap css and js
6.       SQL express 2008